+++ src/Router/Types.fs
@@ -2,87 +2,19 @@ namespace KafkaApplication.Router
 
 open Kafka
 open KafkaApplication
+open Events
 
-//
-// Event data
-//
-
-type SerializedEvent = SerializedEvent of string
-
-type EventToRoute = {
-    Serialized: SerializedEvent
-    Raw: RawEvent
-}
-
-[<RequireQualifiedAccess>]
-module internal EventToRoute =
-    let raw { Raw = raw } = raw
-    let serialized { Serialized = (SerializedEvent serialized) } = serialized
-
-    let parse message =
-        {
-            Serialized = SerializedEvent message
-            Raw = message |> RawEvent.parse
-        }
-
-//
-// Router
-//
+// Errors
 
 type RouterError =
     | StreamNameIsNotInstance of string
 
-type Router = private Router of Map<EventName, StreamName>
-
-module internal Router =
-    open FSharp.Data
-    open System.IO
-    open ServiceIdentification
-
-    type private RoutingSchema = JsonProvider<"src/Router/schema/routingSchema.json">
-
-    let parse path =
-        result {
-            let rawRouting =
-                path
-                |> File.ReadAllText
-                |> RoutingSchema.Parse
-
-            let! routing =
-                rawRouting.Route
-                |> Seq.map (fun route ->
-                    result {
-                        let! topicInstance =
-                            route.TargetStream
-                            |> Instance.parse "-"
-                            |> Result.ofOption (StreamNameIsNotInstance route.TargetStream)
-
-                        return (EventName route.Event, StreamName.Instance topicInstance)
-                    }
-                )
-                |> Seq.toList
-                |> Result.sequence
-
-            return
-                routing
-                |> Map.ofList
-                |> Router
-        }
-
-    let getStreamFor event (Router router) =
-        router
-        |> Map.tryFind event
-
-    let getOutputStreams (Router router) =
-        router
-        |> Map.toList
-        |> List.map snd
-
-// Errors
-
 type RouterConfigurationError =
     | NotFound of string
     | NotSet
+    | MissingRouteEvent
+    | MissingFromDomain
+    | MissingGetCommonEvent
     | OutputBrokerListNotSet
     | RouterError of RouterError
 
@@ -93,10 +25,22 @@ type ContentBasedRouterApplicationError =
 
 // Content-Based Router Application Configuration
 
+type RouterConfiguration = private RouterConfiguration of Map<EventName, StreamName>
+
+type RouteEvent<'InputEvent, 'OutputEvent> = ProcessedBy -> 'InputEvent -> 'OutputEvent
+
+type internal RouteEventHandler<'InputEvent, 'OutputEvent> =
+    | Simple of RouteEvent<'InputEvent, 'OutputEvent>
+    | WithApplication of (PatternRuntimeParts -> RouteEvent<'InputEvent, 'OutputEvent>)
+
 type internal RouterParts<'InputEvent, 'OutputEvent> = {
     Configuration: Configuration<'InputEvent, 'OutputEvent> option
-    RouterConfiguration: Router option
+    RouterConfiguration: RouterConfiguration option
     RouteToBrokerList: BrokerList option
+    RouteEvent: RouteEventHandler<'InputEvent, 'OutputEvent> option
+    FromDomain: FromDomain<'OutputEvent> option
+    CreateCustomValues: CreateCustomValues<'InputEvent, 'OutputEvent> option
+    GetCommonEvent: GetCommonEvent<'InputEvent, 'OutputEvent> option
 }
 
 [<RequireQualifiedAccess>]
@@ -105,13 +49,17 @@ module internal RouterParts =
         Configuration = None
         RouterConfiguration = None
         RouteToBrokerList = None
+        RouteEvent = None
+        FromDomain = None
+        CreateCustomValues = None
+        GetCommonEvent = None
     }
 
 type ContentBasedRouterApplicationConfiguration<'InputEvent, 'OutputEvent> = private ContentBasedRouterApplicationConfiguration of Result<RouterParts<'InputEvent, 'OutputEvent>, ContentBasedRouterApplicationError>
 
 type internal ContentBasedRouterApplicationParts<'InputEvent, 'OutputEvent> = {
     Application: KafkaApplication<'InputEvent, 'OutputEvent>
-    RouterConfiguration: Router
+    RouterConfiguration: RouterConfiguration
 }
 
 type ContentBasedRouterApplication<'InputEvent, 'OutputEvent> = internal ContentBasedRouterApplication of Result<ContentBasedRouterApplicationParts<'InputEvent, 'OutputEvent>, ContentBasedRouterApplicationError>